Remove Premine from Bitcoin Gold
Bitcoin Gold developers added a premine to the blockchain.  This patch is a minimal edit to remove that.

If the majority of Bitcoin Gold GPU miners use this client, there will be no premine on the Bitcoin Gold blockchain.

It is up to the miners to decide which client they want to use to mine.